refactor: CardGame 페이지 폴더 추가
