+++ frontend/src/features/miniGame/cardGame/pages/CardGameReadyPage/CardGameReadyPage.stories.tsx
