+++ frontend/src/features/miniGame/cardGame/pages/CardGameReadyPage/CardGameReadyPage.styled.ts
